SpeechEngine/requirements.txt
django
djangorestframework
redis
speechbrain
django-rest-swagger
drf_yasg